feat: add ClientBody component
